moved transformer with-stmt and dict comprehension tests to separate files
